Add cocktail details page and edit functionality
Added an "Edit" button in `Cocktails.razor` to navigate to a new
`DrinkDetails.razor` page for viewing and editing cocktail details.
The new page fetches cocktail data, including ingredients and
instructions, and displays it with  `MudBlazor` components. Introduced a loading spinner for data fetch when getting data for a drink.